+++ napps_server/api/users.py
@@ -4,8 +4,8 @@
 # Third-party imports
 from flask import Blueprint, jsonify, request
 
-from napps_server.core.decorators import validate_json, validate_schema
-from napps_server.core.decorators import requires_token
+from napps_server.core.decorators import (requires_token, validate_json,
+                                          validate_schema)
 from napps_server.core.exceptions import NappsEntryDoesNotExists
 # Local source tree imports
 from napps_server.core.models import User
@@ -116,13 +116,25 @@ def confirm_user(username, token):
     user.send_welcome()
     return '', 200
 
-@api.route("/users/<username>", methods=['DELETE'])
-@requires_token
+
+@api.route("/users/<username>/", methods=['DELETE'])
 def delete_user(username):
+    """Method used to create a endpoint to delete a user.
+
+    The endpoint created is /users/<username>/
+    """
+    content = request.get_json()
+    token = content.get('token', None)
+
     try:
-       user = User.get(username)
-       user.delete()
+        user = User.get(username)
     except NappsEntryDoesNotExists:
         return jsonify({"error": "User not found."}), 404
+
+    if token != user.token.hash:
+        msg = "The user {} can't be deleted using the token {}."
+        return jsonify({"error": msg.format(user.username, token)}), 404
+
+    user.delete()
     msg = 'The user {} was deleted.'.format(username)
-    return jsonify({'success': msg })
+    return jsonify({'success': msg})
